https://github.com/olofhagsand/cligen/issues/13

+++ cligen_buf.c
@@ -178,7 +178,7 @@ int
 cbuf_append(cbuf       *cb,
             int        c)
 {
-    size_t diff;
+    int diff;
 
     /* make sure we have enough space */
     diff = cb->cb_buflen - (cb->cb_strlen + 1);
